requirements.txt
asgiref==3.8.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-lifecycle==1.2.4
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.0
idna==3.11
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.9.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.6.2
